Optical Properties

RRTMGP.OpticalProps.AbstractOpticalProps — Type

AbstractOpticalProps{FT<:AbstractFloat,I<:Int}

Abstract optical properties

RRTMGP.OpticalProps.AbstractOpticalPropsArry — Type

AbstractOpticalPropsArry{FT,I} <: AbstractOpticalProps{FT,I}

Abstract optical properties array. Includes arrays for optical depth and, depending on the approximation used, other variables.

RRTMGP.OpticalProps.AbstractOpticalPropsPGP — Type

AbstractOpticalPropsPGP{FT,I} <: AbstractOpticalProps{FT,I}

Abstract optical properties array per-grid-point. Includes arrays for optical depth and, depending on the approximation used, other variables.

RRTMGP.OpticalProps.OpticalPropsBase — Type

OpticalPropsBase{FT,I} <: AbstractOpticalProps{FT,I}

Base class for optical properties.

Fields

band2gpt

Map from band to g-point. (g_point_begin, g_point_end) = band2gpt[2,band]

gpt2band

Map from g-point to band. band = gpt2band(gpt)

band_lims_wvn

Wavenumber band limits. (upper_band_wvn, lower_band_wvn) = band_lims_wvn[2,band]

name

Name of optical properties

RRTMGP.OpticalProps.OneScalar — Type

OneScalar{FT,I} <: AbstractOpticalPropsArry{FT,I}

Single scalar approximation for optical depth, used in calculations accounting for extinction and emission

Fields

base

Base optical properties, see OpticalPropsBase

τ

Optical depth

RRTMGP.OpticalProps.TwoStream — Type

TwoStream{FT,I} <: AbstractOpticalPropsArry{FT,I}

Two stream approximation for optical depth, used in calculations accounting for extinction and emission

Fields

base

Base optical properties, see OpticalPropsBase

τ

Optical depth

ssa

Single-scattering albedo

g

Asymmetry parameter

RRTMGP.OpticalProps.OneScalarPGP — Type

OneScalarPGP{FT,I} <: AbstractOpticalPropsArry{FT,I}

Single scalar approximation for optical depth, used in calculations accounting for extinction and emission per-grid-point

Fields

base

Base optical properties, see OpticalPropsBase

τ

Optical depth

RRTMGP.OpticalProps.TwoStreamPGP — Type

TwoStreamPGP{FT,I} <: AbstractOpticalPropsArry{FT,I}

Two stream approximation for optical depth, used in calculations accounting for extinction and emission per-grid-point

Fields

base

Base optical properties, see OpticalPropsBase

τ

Optical depth

ssa

Single-scattering albedo

g

Asymmetry parameter

RRTMGP.OpticalProps.delta_scale! — Function

delta_scale!(this::OneScalar{FT}, fwd_scat_frac)

this optical properties, see OneScalar

fwd_scat_frac forward scattering fraction

delta_scale!(this::TwoStream{FT}, for_ = nothing) where {FT<:AbstractFloat}

this optical properties, see TwoStream

fwd_scat_frac forward scattering fraction

RRTMGP.OpticalProps.validate! — Function

validate!(this::OneScalar{FT}) where FT

Validate values of optical properties

validate!(this::TwoStream{FT}) where FT

Validate values and sizes of optical properties

RRTMGP.OpticalProps.increment! — Function

increment!(opin::AbstractOpticalPropsArry, opio::AbstractOpticalPropsArry)

Increment (combine) two sets of optical properties

RRTMGP.OpticalProps.bands_are_equal — Function

bands_are_equal(this::AbstractOpticalProps{FT}, that::AbstractOpticalProps{FT}) where FT

Boolean that indicates if the bands of two objects the same, (same number, same wavelength limits), given

this optical properties, see AbstractOpticalProps

that optical properties, see AbstractOpticalProps

RRTMGP.OpticalProps.gpoints_are_equal — Function

gpoints_are_equal(this::AbstractOpticalProps{FT}, that::AbstractOpticalProps{FT}) where {FT}

Boolean that indicates if the g-point structure of two objects the same, (same bands, same number of g-points, same mapping between bands and g-points), given

this optical properties, see AbstractOpticalProps

that optical properties, see AbstractOpticalProps