Atmospheric State

RRTMGP.AtmosphericStates.AtmosphericState — Type

AtmosphericState{FT}

Atmospheric conditions, used to compute optical properties

Fields

gas_conc

Gas concentrations, in the form of volume mixing ratios

p_lay

Layer pressures [Pa, mb]; (ncol,nlay)

p_lev

Level pressures [Pa, mb]; (ncol,nlay+1)

t_lay

Layer temperatures [K]; (ncol,nlay)

t_lev

Level temperatures [K]; (ncol,nlay+1)

t_sfc

Surface temperatures [K]; (ncol)

col_gas

Column amounts for each gas, plus col_dry. gas amounts [molec/cm^2]

col_dry

Number of molecules per cm-2 of dry air

tropo

Troposphere mask: itropo = merge(1,2,tropo[icol,ilay]); itropo = 1 lower atmosphere; itropo = 2 upper atmosphere

tropo_lims

Layer limits of upper, lower atmospheres

gt_0_tropo_lims

Any tropospheric limits greater than 0

mesh_orientation

Mesh orientation, see MeshOrientation

nlay

Number of layers.

ncol

Number of columns.

RRTMGP.AtmosphericStates.AtmosphericStatePGP — Type

AtmosphericStatePGP{FT}

Atmospheric conditions per-grid-point, used to compute optical properties

Fields

gas_conc

Gas concentrations, in the form of volume mixing ratios

p_lay

Layer pressures [Pa, mb]

p_lev

Level pressures [Pa, mb]

t_lay

Layer temperatures [K]

t_lev

Level temperatures [K]

t_sfc

Surface temperatures [K]

col_gas

Column amounts for each gas, plus col_dry. gas amounts [molec/cm^2]

col_dry

Number of molecules per cm-2 of dry air

tropo

Troposphere mask: itropo = merge(1,2,tropo); itropo = 1 lower atmosphere; itropo = 2 upper atmosphere

tropo_lims

Layer limits of upper, lower atmospheres

gt_0_tropo_lims

Any tropospheric limits greater than 0

mesh_orientation

Mesh orientation, see MeshOrientation

ilay

Index of i-th layer.

RRTMGP.AtmosphericStates.get_col_dry — Function

get_col_dry(vmr_h2o, plev, tlay, latitude=nothing)

Utility function, provided for user convenience computes column amounts of dry air using hydrostatic equation

input

vmr_h2o volume mixing ratio of water vapor to dry air; (ncol,nlay)

plev Layer boundary pressures [Pa] (ncol,nlay+1)

tlay Layer temperatures [K] (ncol,nlay)

latitude Latitude [degrees] (ncol)

RRTMGP.AtmosphericStates.extrap_lower — Function

extrap_lower(p_lay::Array{FT},
             p_lev::FT,
             ϕ_lay::Array{FT})

Extrapolate lower boundary (weighted by pressure)

RRTMGP.AtmosphericStates.extrap_upper — Function

extrap_upper(p_lay::Array{FT},
             p_lev::FT,
             ϕ_lay::Array{FT}) where {FT<:AbstractFloat,I<:Int}

Extrapolate upper boundary (weighted by pressure)

RRTMGP.AtmosphericStates.interpolate_var — Function

interpolate_var(p_lay::Array{FT},
                p_lev::FT,
                ϕ_lay::Array{FT}) where {FT<:AbstractFloat,I<:Int}

Interpolate variable ϕ (weighted by pressure)