Gas Optics

RRTMGP.GasOptics.GasOpticsAux — Type

GasOpticsAux{I}

Auxiliary indexes for variables in the upper and lower levels of the atmosphere.

Fields

idx_minor

Indexes for determining col_gas

idx_minor_scaling

Indexes that have special treatment in density scaling

RRTMGP.GasOptics.GasOpticsVars — Type

GasOpticsVars{FT,I}

Variables defined at

upper (log(p) < ref.presstroplog)

lower (log(p) > ref.presstroplog)

levels of the atmosphere for both full and reduced sets of gases.

Fields

minor_limits_gpt

Minor g-point limits

minor_scales_with_density

Minor scales with density

scale_by_complement

Scale by complement

kminor_start

Start index for minor gas absorption coefficient

kminor

Absorption coefficient of minor species (n_minor,η,temperature)

scaling_gas

Scaling gas

minor_gases

Minor gases

RRTMGP.GasOptics.AbstractGasOptics — Type

AbstractGasOptics{FT,I} <: AbstractOpticalProps{FT,I}

Abstract gas optics, for long and short wave gas optics dispatch.

RRTMGP.GasOptics.KDistributionLongwave — Type

KDistributionLongwave{FT,I} <: AbstractGasOptics{FT,I}

Gas optics with internal sources (for longwave radiation)

Fields

ref

Reference state data

optical_props

Base optical properties

lower

Gas optics variables in the lower atmosphere, see GasOpticsVars

upper

Gas optics variables in the upper atmosphere, see GasOpticsVars

lower_aux

Auxiliary variables (index maps) in the lower atmosphere, see GasOpticsAux

upper_aux

Auxiliary variables (index maps) in the upper atmosphere, see GasOpticsAux

gas_names

Present gas names

kmajor

Absorption coefficient for major species (g-point,η,pressure,temperature)

flavor

Major species pair; [2, nflav]

gpoint_flavor

Flavor per g-point: lower.flavor = gpoint_flavor[1, 1:ngpt], upper.flavor = gpoint_flavor[2, 1:ngpt]

is_key

Indicates whether a key species is in any band

planck_frac

Stored fraction of Planck irradiance in band for given g-point

totplnk

Integrated Planck irradiance by band; [Planck temperatures,band]

totplnk_delta

Temperature steps in totplnk

krayl

Absorption coefficient for Rayleigh scattering [g-point,η,temperature,upper/lower atmosphere]

RRTMGP.GasOptics.KDistributionShortwave — Type

KDistributionShortwave{FT,I} <: AbstractGasOptics{FT,I}

Gas optics with external sources (for shortwave radiation)

Fields

ref

Reference state data

optical_props

Base optical properties

lower

Gas optics variables in the lower atmosphere, see GasOpticsVars

upper

Gas optics variables in the upper atmosphere, see GasOpticsVars

lower_aux

Auxiliary variables (index maps) in the lower atmosphere, see GasOpticsAux

upper_aux

Auxiliary variables (index maps) in the upper atmosphere, see GasOpticsAux

gas_names

Present gas names

kmajor

Absorption coefficient for major species (g-point,η,pressure,temperature)

flavor

Major species pair; [2, nflav]

gpoint_flavor

Flavor per g-point: lower.flavor = gpoint_flavor[1, 1:ngpt], upper.flavor = gpoint_flavor[2, 1:ngpt]

is_key

Indicates whether a key species is in any band

solar_src

Incoming solar irradiance (g-point)

krayl

Absorption coefficient for Rayleigh scattering (g-point,η,temperature,upper/lower atmosphere)

RRTMGP.GasOptics.InterpolationCoefficients — Type

InterpolationCoefficients{FT,I}

Interpolation coefficients

Fields

jtemp

index for temperature

jpress

index for pressure

j_η

index for binary species parameter (η)

fmajor

fractions for major species

fminor

fractions for minor species. [reference η level (temperature dependent), reference temperature level, flavor, layer]

col_mix

combination of major specie's column amounts

RRTMGP.GasOptics.InterpolationCoefficientsPGP — Type

InterpolationCoefficientsPGP{FT,I}

Interpolation coefficients per grid point

Fields

jtemp

index for temperature

jpress

index for pressure

j_η

index for binary species parameter (η)

fmajor

fractions for major species

fminor

fractions for minor species. [reference η level (temperature dependent), reference temperature level, flavor, layer]

col_mix

combination of major specie's column amounts

RRTMGP.GasOptics.gas_optics! — Function

gas_optics!(this::KDistributionLongwave{FT,I},
            as::AtmosphericState{FT,I},
            optical_props::AbstractOpticalPropsArry{FT,I},
            sources::SourceFuncLongWave{FT,I}) where {FT<:AbstractFloat,I<:Int}

Compute gas optical depth and Planck source functions given:

this gas optics, see KDistributionLongwave

as atmospheric state, see AtmosphericState

optical_props optical properties, see AbstractOpticalPropsArry

sources longwave sources, see SourceFuncLongWave

gas_optics!(this::KDistributionShortwave{FT,I},
            as::AtmosphericState{FT,I},
            optical_props::AbstractOpticalPropsArry{FT,I},
            last_call=false) where {FT<:AbstractFloat,I<:Int}

Compute gas optical depth given:

this gas optics, see KDistributionShortwave

as atmospheric state, see AtmosphericState

optical_props optical properties, see AbstractOpticalPropsArry

RRTMGP.GasOptics.get_k_dist_lw — Function

get_k_dist_lw(totplnk,
              planck_frac,
              rayl_lower,
              rayl_upper,
              ref::ReferenceState{FT},
              args...) where {FT<:AbstractFloat}

Initialize object based on data read from netCDF file however the user desires. Rayleigh scattering tables may or may not be present; this is indicated with allocation status This interface is for the internal-sources object – includes Planck functions and fractions

RRTMGP.GasOptics.get_k_dist_sw — Function

get_k_dist_sw(solar_src,
              rayl_lower,
              rayl_upper,
              ref::ReferenceState{FT},
              args...) where {FT<:AbstractFloat}

Initialize object based on data read from netCDF file however the user desires. Rayleigh scattering tables may or may not be present; this is indicated with allocation status This interface is for the external-sources object – includes TOA source function table