Source Functions

RRTMGP.GraySources.GraySourceLWNoScat — Type

GraySourceLWNoScat{
FT<:AbstractFloat,
FTA2D<:AbstractArray{FT,2},

} <: AbstractGraySource{FT}

Longwave sources: computed at layer center, layer edges, and at the surface for no scattering calculations

Fields

lay_source

Planck source at layer average temperature [W/m2] (nlay, ncol, ngpt)

lev_source_inc

Planck source at layer edge in increasing ilay direction [W/m2] (nlay+1, ncol, ngpt), includes spectral weighting that accounts for state-dependent frequency to g-space mapping

lev_source_dec

Planck source at layer edge in decreasing ilay direction [W/m2] (nlay+1, ncol, ngpt), includes spectral weighting that accounts for state-dependent frequency to g-space mapping

sfc_source

Surface source

source_up

temporary storage array used in no scattering GPU calculations

source_dn

temporary storage array used in no scattering GPU calculations

RRTMGP.GraySources.GraySourceLW2Str — Type

GraySourceLW2Str{
FT<:AbstractFloat,
FTA2D<:AbstractArray{FT,2},

} <: AbstractGraySource{FT}

Longwave sources: computed at layer center, layer edges, and at the surface for 2-stream calculations

Fields

lay_source

Planck source at layer average temperature [W/m2] (nlay, ncol, ngpt)

lev_source_inc

Planck source at layer edge in increasing ilay direction [W/m2] (nlay+1, ncol, ngpt), includes spectral weighting that accounts for state-dependent frequency to g-space mapping

lev_source_dec

Planck source at layer edge in decreasing ilay direction [W/m2] (nlay+1, ncol, ngpt), includes spectral weighting that accounts for state-dependent frequency to g-space mapping

sfc_source

Surface source

lev_source

level source [W/m2] (nlay+1, ncol, ngpt), used in 2 stream calculations

src_up

temporary storage array, used in 2 stream calculations

src_dn

temporary storage array, used in 2 stream calculations

Rdif

temporary storage array, used in 2 stream calculations

Tdif

temporary storage array, used in 2 stream calculations

albedo

temporary storage array, used in 2 stream calculations

src

temporary storage array, used in 2 stream calculations