gas optics

Gas optics RRTMG

JRRTMGP.mo_gas_optics_rrtmgp.gas_optics! — Function.

gas_optics!(this::ty_gas_optics_rrtmgp,
                   play,
                   plev,
                   tlay,
                   tsfc,
                   gas_desc::ty_gas_concs,
                   optical_props::ty_optical_props_arry,
                   sources::ty_source_func_lw,
                   col_dry=nothing,
                   tlev=nothing)

Compute gas optical depth and Planck source functions, given temperature, pressure, and composition

Inputs

this (ty_gas_optics_rrtmgp structure)

play[ncol,nlay] - layer pressures [Pa, mb]

plev[ncol,nlay+1] - level pressures [Pa, mb]

tlay[ncol,nlay] - layer temperatures [K]

tsfc[ncol] - surface skin temperatures [K]

gas_desc - gas volume mixing ratios (tygasconcs structure)

Optional inputs

col_dry[ncol,nlay] - column dry amount

tlev[ncol,nlay+1] - level temperatures [K]

Output

optical_props - (ty_optical_props_arry structure) Optical properties

sources - (ty_source_func_lw) Planck sources

source

gas_optics!(this::ty_gas_optics_rrtmgp,
                   play,
                   plev,
                   tlay,
                   gas_desc::ty_gas_concs,    # mandatory inputs
                   optical_props::ty_optical_props_arry,
                   toa_src,        # mandatory outputs
                   col_dry=nothing)

Compute gas optical depth given temperature, pressure, and composition

Inputs

this - (ty_gas_optics_rrtmgp structure)

play[ncol,nlay] - layer pressures [Pa, mb]

plev[ncol,nlay+1] - level pressures [Pa, mb]

tlay[ncol,nlay] - layer temperatures [K]

gas_desc - (ty_gas_concs structure) gas volume mixing ratios

Optional inputs

col_dry[ncol,nlay] - column dry amount

Outputs

optical_props - (ty_optical_props_arry structure)

toa_src[ncol,ngpt] - incoming solar irradiance

source

JRRTMGP.mo_gas_optics_rrtmgp.source — Function.

source(this::ty_gas_optics_rrtmgp,
              ncol, nlay, nbnd, ngpt,
              play, plev, tlay, tsfc,
              jtemp, jpress, jeta, tropo, fmajor,
              sources::ty_source_func_lw,
              tlev)

Compute Planck source functions at layer centers and levels

Inputs

this - (structure ty_gas_optics_rrtmgp)

ncol - number of columns

nlay - number of layers

nbnd - number of bands

ngpt - number of g-points

play[ncol,nlay] - layer pressures [Pa, mb]

plev[ncol,nlay+1] - level pressures [Pa, mb]

tlay[ncol,nlay] - layer temperatures [K]

tsfc[ncol] - surface skin temperatures [K]

jtemp[ncol,nlay] - interpolation index for temperature

jpress[ncol,nlay] - interpolation index for pressure

tropo[ncol,nlay]

fmajor[2,2,2,get_nflav(this),ncol,nlay)]

jeta[2,get_nflav(this),ncol,nlay] - interpolation index for mixing ratio

sources - (structure ty_source_func_lw) - Planck sources

Optional Inputs

tlev[ncol,nlay+1] - level temperatures [K]

Outputs

sources - (structure ty_source_func_lw)

source

JRRTMGP.mo_gas_optics_rrtmgp.init_abs_coeffs — Function.

init_abs_coeffs(rayl_lower, rayl_upper,
available_gases::ty_gas_concs,
gas_names,
key_species,
band2gpt,
band_lims_wavenum,
press_ref,
press_ref_trop,
temp_ref,
temp_ref_p,
temp_ref_t,
vmr_ref,
kmajor,
kminor_lower,
kminor_upper,
gas_minor,
identifier_minor,
minor_gases_lower,
minor_gases_upper,
minor_limits_gpt_lower,
minor_limits_gpt_upper,
minor_scales_with_density_lower,
minor_scales_with_density_upper,
scaling_gas_lower,
scaling_gas_upper,
scale_by_complement_lower,
scale_by_complement_upper,
kminor_start_lower,
kminor_start_upper)

Initialize absorption coefficient arrays, including Rayleigh scattering tables if provided (allocated)

Inputs

available_gases - (structure ty_gas_concs)

gas_names - names of major gases

key_species[:, :, :] - Key species pair for each band

band2gpt[:, :]

band_lims_wavenum[:, :]

press_ref[:]

temp_ref[:]

press_ref_trop

temp_ref_p

temp_ref_t

vmr_ref[:, :, :]

kmajor[:, :, :, :]

kminor_lower[:, :, :]

kminor_upper[:, :, :]

gas_minor - names of minor gases

identifier_minor - identifier for minor gases

minor_gases_lower - minor gases in the lower atmosphere

minor_gases_upper - minor gases in the upper atmosphere

minor_limits_gpt_lower[:, :]

minor_limits_gpt_upper[:, :]

minor_scales_with_density_lower[:]

minor_scales_with_density_upper[:]

scaling_gas_lower

scaling_gas_upper

scale_by_complement_lower

scale_by_complement_upper

kminor_start_lower[:]

kminor_start_upper[:]

rayl_lower[:, :, :]

rayl_upper[:, :, :]

Outputs

this - (structure ty_gas_optics_rrtmgp)

source