Optical Properties

RRTMGP.OpticalProps.OpticalPropsBase — Type

OpticalPropsBase{FT,I} <: AbstractOpticalProps{FT,I}

Base class for optical properties.

Fields

band2gpt

Map from band to g-point. (begin g-point, end g-point) = band2gpt(2,band)

gpt2band

Map from g-point to band. band = gpt2band(g-point)

band_lims_wvn

wavenumber band limits. (upper and lower wavenumber by band) = bandlimswvn(2,band)

name

Name of optical properties

RRTMGP.OpticalProps.OneScalar — Type

OneScalar{FT,I} <: AbstractOpticalPropsArry{FT,I}

Single scalar approximation for optical depth, used in calculations accounting for extinction and emission

Fields

base

Base optical properties, see

τ

optical depth

RRTMGP.OpticalProps.TwoStream — Type

TwoStream{FT,I} <: AbstractOpticalPropsArry{FT,I}

Two stream approximation for optical depth, used in calculations accounting for extinction and emission

Fields

base

optical properties discretization and mapping

τ

optical depth

ssa

single-scattering albedo

g

asymmetry parameter

RRTMGP.OpticalProps.delta_scale! — Function

delta_scale!(this::OneScalar{FT}, fwd_scat_frac)

this optical properties, see OneScalar

fwd_scat_frac forward scattering fraction

delta_scale!(this::TwoStream{FT}, for_ = nothing) where {FT<:AbstractFloat}

this optical properties, see TwoStream

fwd_scat_frac forward scattering fraction

RRTMGP.OpticalProps.validate! — Function

validate!(this::OneScalar{FT}) where FT

validate!(this::TwoStream{FT}) where FT

RRTMGP.OpticalProps.increment! — Function

increment!(opin::AbstractOpticalPropsArry, opio::AbstractOpticalPropsArry)

RRTMGP.OpticalProps.bands_are_equal — Function

bands_are_equal(this::AbstractOpticalProps{FT}, that::AbstractOpticalProps{FT}) where FT

Boolean that indicates if the bands of two objects the same, (same number, same wavelength limits), given

this optical properties, see AbstractOpticalProps

that optical properties, see AbstractOpticalProps

RRTMGP.OpticalProps.gpoints_are_equal — Function

gpoints_are_equal(this::AbstractOpticalProps{FT}, that::AbstractOpticalProps{FT}) where {FT}

Boolean that indicates if the g-point structure of two objects the same, (same bands, same number of g-points, same mapping between bands and g-points), given

this optical properties, see AbstractOpticalProps

that optical properties, see AbstractOpticalProps