Lookup tables

RRTMGP.LookUpTables.LookUpMinor — Type

LookUpMinor{O, G, K}

Lookup table for computing optical properties of minor gases.

Fields

bnd_st: Starting index to idx_gases_minor_lower/upper for each band (n_bnd + 1).

gpt_st: Starting index in kminor_lower/upper for each g-point (n_gpt + 1).

gasdata: Indices for minor gases contributing to absorption in the lower/upper atmosphere, indices for scaling gases, whether the minor gas scales with density, and whether the minor gas scales by complement; (4, n_min_absrb_lower/upper).

kminor: Minor absorption coefficient in the lower/upper atmosphere (n_η, n_t_ref, n_contrib_lower/upper).

RRTMGP.LookUpTables.ReferencePoints — Type

ReferencePoints{RD1, RD3}

Log of reference pressures, reference temperatures, and volume mixing ratios used by the lookup table.

Fields

ln_p_ref: Log of reference pressures used by the lookup table (n_p_ref).

t_ref: Reference temperatures used by the lookup table (n_t_ref).

vmr_ref: Reference volume mixing ratios used by the lookup table (2, n_gases, n_t_ref).

RRTMGP.LookUpTables.LookUpPlanck — Type

LookUpPlanck{PD1, PD2, PD4}

Lookup data for Planck source calculations.

Fields

planck_fraction: Planck fraction (n_η, n_p_ref, n_t_ref, n_gpt).

t_planck: Reference temperatures for Planck source calculations (n_t_plnk).

tot_planck: Total Planck source for each band (n_t_plnk, n_bnd).

RRTMGP.LookUpTables.BandData — Type

BandData{BNDI1, BNDI2, BNDD2}

Band/g-point data for the lookup table.

Fields

major_gpt2bnd: Map from g-point to band.

bnd_lims_gpt: Starting and ending g-point for each band (2, n_bnd).

bnd_lims_wn: Starting and ending wavenumber for each band (2, n_bnd).

RRTMGP.LookUpTables.LookUpLW — Type

LookUpLW{FT, IA3D, FTA4D, BND, P, R, LMNR} <: AbstractLookUp

Longwave lookup tables, used to compute optical properties.

Fields

idx_h2o: Vmr array index for H₂O.

p_ref_tropo: Reference pressure separating the upper and lower atmosphere.

p_ref_min: Minimum pressure supported by RRTMGP lookup tables.

t_ref_min: Minimum temperature supported by RRTMGP lookup tables (first reference temperature).

t_ref_max: Maximum temperature supported by RRTMGP lookup tables (last reference temperature).

key_species: Major absorbing species in each band (2, n_atmos_layers, n_bnd).

kmajor: Major absorption coefficient (n_η, n_p_ref + 1, n_t_ref, n_gpt).

planck: Lookup data for Planck source calculations.

band_data: Band data.

ref_points: Reference temperatures, pressures and volume mixing ratios.

minor_lower: Lookup data for minor gases in the lower atmosphere.

minor_upper: Lookup data for minor gases in the upper atmosphere.

RRTMGP.LookUpTables.LookUpSW — Type

LookUpSW{FT, IA3D, FTA1D, FTA3D, FTA4D, BND, R, LMNR} <: AbstractLookUp

Shortwave lookup tables, used to compute optical properties.

Fields

idx_h2o: Vmr array index for H₂O.

p_ref_tropo: Reference pressure separating the upper and lower atmosphere.

p_ref_min: Minimum pressure supported by RRTMGP lookup tables.

t_ref_min: Minimum temperature supported by RRTMGP lookup tables (first reference temperature).

t_ref_max: Maximum temperature supported by RRTMGP lookup tables (last reference temperature).

solar_src_tot: Total solar irradiation.

key_species: Major absorbing species in each band (2, n_atmos_layers, n_bnd).

kmajor: Major absorption coefficient (n_η, n_p_ref + 1, n_t_ref, n_gpt).

band_data: Band data.

ref_points: Reference temperatures, pressures and volume mixing ratios.

rayl_lower: Rayleigh absorption coefficient for the lower atmosphere (n_η, n_t_ref, n_gpt).

rayl_upper: Rayleigh absorption coefficient for the upper atmosphere (n_η, n_t_ref, n_gpt).

solar_src_scaled: Relative solar source contribution from each g-point (n_gpt).

minor_lower: Lookup data for minor gases in the lower atmosphere.

minor_upper: Lookup data for minor gases in the upper atmosphere.

RRTMGP.LookUpTables.LookUpCld — Type

LookUpCld{D, B, L, I, W} <: AbstractLookUp

Lookup table for cloud optics.

This struct stores the lookup tables for determining the extinction coefficient, single-scattering albedo, and asymmetry parameter g as a function of effective radius. We compute the optical depth tau (= extinction coefficient * condensed water path) and the products taussa and taussa*g for liquid and ice cloud separately. These are used to determine the optical properties of ice and water cloud together.

Fields

dims: Dimensions nband, nrghice, nsize_liq, nsize_ice, pair.

bounds: Particle size lower and upper bounds and factor for LUT interpolation for liquid and ice particles.

liqdata: Liquid extinction coefficient, single scattering albedo and asymmetry parameter (3*nsize_liq, nbnd).

icedata: Ice extinction coefficient, single scattering albedo and asymmetry parameter (3*nsize_ice, nbnd, nrghice).

bnd_lims_wn: Beginning and ending wavenumber for each band [cm⁻¹] (2, nband).

RRTMGP.LookUpTables.LookUpAerosolMerra — Type

LookUpAerosolMerra{D, D1, D2, D3, D4, W} <: AbstractLookUp

MERRA lookup table for aerosols.

This struct stores the lookup tables for determining the extinction coefficient, single-scattering albedo, and asymmetry parameter g as a function of aerosol particle size, relative humidity and band. Data is provided for dust, sea salt, sulfate, black carbon (hydrophobic and hydrophilic) and organic carbon (hydrophobic and hydrophilic).

Fields

dims: Dimensions nband, nval, nbin, nrh, pair.

size_bin_limits: Beginning and ending limit for each MERRA aerosol size bin [microns].

rh_levels: Relative humidity levels for MERRA hydrophilic aerosols.

dust: Dust (nval, nbin, nband).

sea_salt: Sea salt (nval, nrh, nbin, nband).

sulfate: Sulfate (nval, nrh, nband).

black_carbon_rh: Black carbon, hydrophilic (nval, nrh, nband).

black_carbon: Black carbon, hydrophobic (nval, nband).

organic_carbon_rh: Organic carbon, hydrophilic (nval, nrh, nband).

organic_carbon: Organic carbon, hydrophobic (nval, nband).

bnd_lims_wn: Beginning and ending wavenumber for each band [cm⁻¹] (2, nband).

iband_550nm: Band number index corresponding to 550 nm.

RRTMGP.LookUpTables.AbstractLookUp — Type

AbstractLookUp

Abstract lookup table for longwave and shortwave problems.