Source Functions

RRTMGP.Sources.AbstractSourceLW — Type

AbstractSourceLW{FT,FTA1D,FTA2D}

Abstract longwave source for no-scattering and two stream longwave solvers.

RRTMGP.Sources.SourceLWNoScat — Type

SourceLWNoScat{FT,FTA1D,FTA2D} <: AbstractSourceLW{FT,FTA1D,FTA2D}

Longwave sources: computed at layer center, layer edges, and at the surface for no scattering calculations

Fields

param_set: Parameter set

lay_source: Planck source at layer average temperature [W/m2] (nlay, ncol)

lev_source_inc: Planck source at layer edge in increasing ilay direction [W/m2] (nlay, ncol), includes spectral weighting that accounts for state-dependent frequency to g-space mapping

lev_source_dec: Planck source at layer edge in decreasing ilay direction [W/m2] (nlay, ncol), includes spectral weighting that accounts for state-dependent frequency to g-space mapping

sfc_source: Surface source [W/m2] (ncol)

src_up: temporary storage array used in no scattering GPU calculations (nlay, ncol)

src_dn: temporary storage array used in no scattering GPU calculations (nlay, ncol)

RRTMGP.Sources.SourceLW2Str — Type

SourceLW2Str{FT,FTA1D,FTA2D} <: AbstractSourceLW{FT,FTA1D,FTA2D}

Longwave sources: computed at layer center, layer edges, and at the surface for 2-stream calculations

Fields

param_set: Parameter set

lay_source: Planck source at layer average temperature [W/m2] (nlay, ncol)

lev_source_inc: Planck source at layer edge in increasing ilay direction [W/m2] (nlay, ncol), includes spectral weighting that accounts for state-dependent frequency to g-space mapping

lev_source_dec: Planck source at layer edge in decreasing ilay direction [W/m2] (nlay, ncol), includes spectral weighting that accounts for state-dependent frequency to g-space mapping

sfc_source: Surface source [W/m2] (ncol)

lev_source: level source [W/m2] (nlay+1, ncol), used in 2 stream calculations

src_up: temporary storage array, used in 2 stream calculations (nlay, ncol)

src_dn: temporary storage array, used in 2 stream calculations (nlay, ncol)

Rdif: temporary storage array, used in 2 stream calculations (nlay, ncol)

Tdif: temporary storage array, used in 2 stream calculations (nlay, ncol)

albedo: temporary storage array, used in 2 stream calculations (nlay + 1, ncol)

src: temporary storage array, used in 2 stream calculations (nlay + 1, ncol)

RRTMGP.Sources.SourceSW2Str — Type

SourceSW2Str{FT,FTA1D,FTA2D}

Shortwave sources: computed at layer center, layer edges, and at the surface for 2-stream calculations

Fields

sfc_source: surface source (ncol)

albedo: albedo (nlay + 1, ncol)

src_up: temporary storage array, used in 2 stream calculations (nlay, ncol)

src_dn: temporary storage array, used in 2 stream calculations (nlay, ncol)

Rdif: temporary storage array, used in 2 stream calculations (nlay, ncol)

Tdif: temporary storage array, used in 2 stream calculations (nlay, ncol)

Rdir: temporary storage array, used in 2 stream calculations (nlay, ncol)

Tdir: temporary storage array, used in 2 stream calculations (nlay, ncol)

Tnoscat: temporary storage array, used in 2 stream calculations (nlay, ncol)

src: temporary storage array, used in 2 stream calculations (nlay + 1, ncol)

RRTMGP.Sources.source_func_longwave — Function

source_func_longwave(
    param_set,
    ::Type{FT},
    ncol::Int,
    nlay::Int,
    OPC::Symbol,
    ::Type{DA},
) where {FT<:AbstractFloat,DA}

Initializes the longwave source for one scalar and two stream simulations.

RRTMGP.Sources.source_func_shortwave — Function

source_func_shortwave(
    ::Type{FT},
    ncol::Int,
    nlay::Int,
    opc::Symbol,
    ::Type{DA},
) where {FT<:AbstractFloat,DA}

Initializes the shortwave source for one scalar and two stream simulations.