Volume Mixing Ratios

RRTMGP.Vmrs.AbstractVmr — Type

AbstractVmr{FT}

RRTMGP.Vmrs.Vmr — Type

Vmr{FT,FTA3D} <: AbstractVmr{FT}

Volume mixing ratios for various gases in the atmosphere. This struct can be used when concentrations vary spatially for all gases.

Fields

vmr: volume mixing ratio of all gases as a function of location and column

RRTMGP.Vmrs.VmrGM — Type

VmrGM{FT,FTA1D,FTA2D} <: AbstractVmr{FT}

Volume mixing ratios for various gases in the atmosphere. This struct can be used when only H₂O and O₃ concentrations vary spatially and a global mean is used for all other gases.

Fields

vmr_h2o: volume mixing ratio of H₂O

vmr_o3: volume mixing ratio of Ozone

vmr: volume mixing ratio of all other gases, which are independent of location and column

RRTMGP.Vmrs.init_vmr — Function

init_vmr(
    ngas,
    nlay,
    ncol,
    ::Type{FT},
    ::Type{DA};
    gm::Bool = true,
) where {FT<:AbstractFloat,DA}

Initialize the Vmr struct.

RRTMGP.Vmrs.get_vmr — Function

get_vmr(
    vmr::VmrGM{FT},
    ig::Int,
    ilay::Int,
    icol::Int,
) where {FT<:AbstractFloat}

Obtain volume mixing ratio of gas ig for layer ilay of column icol.

get_vmr(
    vmr::Vmr{FT},
    ig::Int,
    ilay::Int,
    icol::Int,
) where {FT<:AbstractFloat}

Obtain volume mixing ratio of gas ig for layer ilay of column icol.