Gas Optics

Gas Optics

RRTMGP.GasOptics.GasOpticsAux — Type.

GasOpticsAux{I}

Auxiliary indexes for variables in the upper and lower levels of the atmosphere.

Fields

idx_minor

indexes for determining col_gas

idx_minor_scaling

indexes that have special treatment in density scaling

source

RRTMGP.GasOptics.GasOpticsVars — Type.

GasOpticsVars{FT,I}

Variables defined at

upper (log(p) < ref.presstroplog)

lower (log(p) > ref.presstroplog)

levels of the atmosphere for both full and reduced sets of gases.

Fields

minor_limits_gpt

minor g-point limits

minor_scales_with_density

minor scales with density

scale_by_complement

scale by complement

kminor_start

kminor start

kminor

absorption coefficient of minor species

scaling_gas

scaling gas

minor_gases

minor gases

source

RRTMGP.GasOptics.KDistributionLongwave — Type.

KDistributionLongwave{FT,I} <: AbstractGasOptics{FT,I}

Gas optics with internal sources (for longwave radiation)

Fields

ref

Reference state data

optical_props

Base optical properties

lower

GasOpticsVars in the lower atmosphere

upper

GasOpticsVars in the upper atmosphere

lower_aux

Auxiliary variables (index maps) in the lower atmosphere

upper_aux

Auxiliary variables (index maps) in the upper atmosphere

gas_names

Present gas names

i_gases

Indexes of present gas names

kmajor

Absorption coefficient for major species (g-point,η,pressure,temperature)

flavor

major species pair; [2, nflav]

gpoint_flavor

Flavor per g-point: lower.flavor = gpoint_flavor[1, 1:ngpt], upper.flavor = gpoint_flavor[2, 1:ngpt]

is_key

Indicates whether a key species is in any band

planck_frac

Stored fraction of Planck irradiance in band for given g-point

totplnk

integrated Planck irradiance by band; [Planck temperatures,band]

totplnk_delta

temperature steps in totplnk

krayl

Absorption coefficient for Rayleigh scattering [g-point,η,temperature,upper/lower atmosphere]

source

RRTMGP.GasOptics.KDistributionShortwave — Type.

KDistributionShortwave{FT,I} <: AbstractGasOptics{FT,I}

Gas optics with external sources (for shortwave radiation)

Fields

ref

Reference state data

optical_props

Base optical properties

lower

GasOpticsVars in the lower atmosphere

upper

GasOpticsVars in the upper atmosphere

lower_aux

Auxiliary variables (index maps) in the lower atmosphere

upper_aux

Auxiliary variables (index maps) in the upper atmosphere

gas_names

Present gas names

i_gases

Indexes of present gas names

kmajor

Absorption coefficient for major species (g-point,η,pressure,temperature)

flavor

major species pair; [2, nflav]

gpoint_flavor

Flavor per g-point: lower.flavor = gpoint_flavor[1, 1:ngpt], upper.flavor = gpoint_flavor[2, 1:ngpt]

is_key

Indicates whether a key species is in any band

solar_src

incoming solar irradiance (g-point)

krayl

absorption coefficient for Rayleigh scattering (g-point,η,temperature,upper/lower atmosphere)

source