Source Functions

RRTMGP.SourceFunctions.AbstractSourceFunc — Type

AbstractSourceFunc{FT,I} <: AbstractOpticalProps{FT, I}

Abstract source function, to dispatch between short-wave and long-wave

RRTMGP.SourceFunctions.SourceFuncLongWave — Type

SourceFuncLongWave{FT, I} <: AbstractOpticalProps{FT, I}

Longwave sources: computed at layer center, at layer edges using spectral mapping in each direction separately, and at the surface

Fields

optical_props

optical properties, see OpticalPropsBase

lay_source

Planck source at layer average temperature [W/m2] (ncol, nlay, ngpt)

lev_source_inc

Planck source at layer edge in increasing ilay direction [W/m2] (ncol, nlay+1, ngpt), includes spectral weighting that accounts for state-dependent frequency to g-space mapping

lev_source_dec

Planck source at layer edge in decreasing ilay direction [W/m2] (ncol, nlay+1, ngpt), includes spectral weighting that accounts for state-dependent frequency to g-space mapping

sfc_source

Surface source

p_frac

Temporary array for Planck fraction

RRTMGP.SourceFunctions.SourceFuncLongWavePGP — Type

SourceFuncLongWavePGP{FT, I} <: OpticalPropsBase{FT, I}

Longwave sources: computed at layer center, at layer edges using spectral mapping in each direction separately, and at the surface per grid point

Fields

optical_props

optical properties, see OpticalPropsBase

lay_source

Planck source at layer average temperature [W/m2] (ngpt)

lev_source_inc

Planck source at layer edge in increasing ilay direction [W/m2] (ngpt), includes spectral weighting that accounts for state-dependent frequency to g-space mapping

lev_source_dec

Planck source at layer edge in decreasing ilay direction [W/m2] (ngpt), includes spectral weighting that accounts for state-dependent frequency to g-space mapping

sfc_source

Surface source

p_frac

Temporary array for Planck fraction

RRTMGP.SourceFunctions.SourceFuncShortWave — Type

SourceFuncShortWave{FT, I} <: AbstractOpticalProps{FT, I}

Shortwave sources

Fields

optical_props

optical properties, see OpticalPropsBase

toa_source

top of atmosphere source

lev_source_inc

Source at layer edge in increasing ilay direction

lev_source_dec

Source at layer edge in decreasing ilay direction