RTE Solver

RTE Solver

RRTMGP.RTESolver.rte_sw! — Function.

rte_sw!(atmos::AbstractOpticalPropsArry,
        top_at_1::Bool,
        μ_0::Array{FT},
        inc_flux,
        sfc_alb_dir,
        sfc_alb_dif,
        fluxes,
        inc_flux_dif=nothing) where {FT<:AbstractFloat}

Compute fluxes fluxes (AbstractFluxes), given

atmos Optical properties provided as arrays (AbstractOpticalPropsArry)

top_at_1 bool indicating if the top of the domain is at index 1 (if not, ordering is bottom-to-top)

μ_0 cosine of solar zenith angle (ncol)

inc_flux incident flux at top of domain W/m2

sfc_alb_dir surface albedo for direct and

sfc_alb_dif diffuse radiation (nband, ncol)

and, optionally,

inc_flux_dif incident diffuse flux at top of domain W/m2

source

RRTMGP.RTESolver.rte_lw! — Function.

rte_lw!(optical_props::AbstractOpticalPropsArry{FT},
        top_at_1,
        sources,
        sfc_emis,
        fluxes,
        inc_flux=nothing,
        n_gauss_angles=nothing) where FT

Interface using only optical properties and source functions as inputs; fluxes as outputs.

optical_props optical properties

top_at_1 indicates that the top of the domain is at index 1

sources radiation sources

sfc_emis emissivity at surface (nband, ncol)

fluxes Array of AbstractFluxes. Default computes broadband fluxes at all levels if output arrays are defined. Can be extended per user desires.

[inc_flux] incident flux at domain top W/m2

[n_gauss_angles] Number of angles used in Gaussian quadrature (no-scattering solution)

source

RRTMGP.RTESolver.expand_and_transpose — Function.

expand_and_transpose(ops::AbstractOpticalProps,arr_in::Array{FT}) where FT

Expand from band to g-point dimension, transpose dimensions (nband, ncol) -> (ncol,ngpt), of arr_out, given

ops - a AbstractOpticalProps

arr_in - input array

source