RTE Solver

RTE Solver

RRTMGP.RTESolver.rte_sw! — Function.

rte_sw!(atmos::AbstractOpticalPropsArry{FT},
        top_at_1::Bool,
        μ_0::Array{FT},
        bcs::ShortwaveBCs{FT},
        fluxes::FluxesBroadBand{FT}) where {FT<:AbstractFloat}

Computes

fluxes broadband fluxes, see AbstractFluxes

given

atmos Optical properties provided as arrays (AbstractOpticalPropsArry)

top_at_1 indicates that the top of the domain is at index 1

μ_0 cosine of solar zenith angle (ncol)

bcs boundary conditions, see ShortwaveBCs

source

RRTMGP.RTESolver.rte_lw! — Function.

rte_lw!(optical_props::AbstractOpticalPropsArry{FT},
        top_at_1::Bool,
        sources::SourceFuncLongWave{FT, I},
        bcs::LongwaveBCs{FT},
        fluxes::FluxesBroadBand{FT},
        n_gauss_angles=nothing) where {FT<:AbstractFloat,I<:Int}

Interface using only optical properties and source functions as inputs; fluxes as outputs.

optical_props optical properties

top_at_1 indicates that the top of the domain is at index 1

sources radiation sources

bcs boundary conditions, see LongwaveBCs

fluxes broadband fluxes, see AbstractFluxes

n_gauss_angles Number of angles used in Gaussian quadrature (no-scattering solution)

source

RRTMGP.RTESolver.expand_and_transpose — Function.

expand_and_transpose(ops::AbstractOpticalProps,arr_in::Array{FT}) where FT

Expand from band to g-point dimension, transpose dimensions (nband, ncol) -> (ncol,ngpt), of arr_out, given

ops - a AbstractOpticalProps

arr_in - input array

source