Fluxes

RRTMGP.Fluxes.FluxLW — Type

FluxLW{FT,FTA2D}

Upward, downward and net longwave fluxes at each level.

Fields

flux_up: upward flux [W/m²] (nlev,ncol)

flux_dn: downward flux [W/m²] (nlev,ncol)

flux_net: net flux [W/m²] (nlev,ncol)

RRTMGP.Fluxes.FluxSW — Type

FluxSW{FT,FTA2D}

Upward, downward and net shortwave fluxes at each level.

Fields

flux_up: upward flux [W/m²] (nlev,ncol)

flux_dn: downward flux [W/m²] (nlev,ncol)

flux_net: net flux [W/m²] (nlev,ncol)

flux_dn_dir: direct downward flux [W/m²] (nlev,ncol)

RRTMGP.Fluxes.set_flux_to_zero! — Function

set_flux_to_zero!(flux::FluxLW{FT}) where {FT<:AbstractFloat}

Set longwave flux to zero

set_flux_to_zero!(flux::FluxLW{FT}, gcol::Int) where {FT<:AbstractFloat}

Set longwave flux for column gcol to zero

set_flux_to_zero!(flux::FluxSW{FT}) where {FT<:AbstractFloat}

Set shortwave flux to zero

set_flux_to_zero!(flux::FluxSW{FT}, gcol::Int) where {FT<:AbstractFloat}

Set shortwave flux for column gcol to zero

RRTMGP.Fluxes.add_to_flux! — Function

add_to_flux!(flux1::FluxLW, flux2::FluxLW)

add longwave flux2 to longwave flux1 flux1 .+= flux2

add_to_flux!(flux1::FluxLW, flux2::FluxLW, gcol::Int)

add longwave flux2 to longwave flux1 for column gcol flux1 .+= flux2

add_to_flux!(flux1::FluxSW, flux2::FluxSW)

add shortwave flux2 to shortwave flux1 flux1 .+= flux2

add_to_flux!(flux1::FluxSW, flux2::FluxSW, gcol::Int)

add shortwave flux2 to shortwave flux1 for column gcol flux1 .+= flux2