Lookup tables

RRTMGP.LookUpTables.LookUpMinor — Type

LookUpMinor{O, G, K}

Lookup table for computing optical properties of minor gases.

RRTMGP.LookUpTables.ReferencePoints — Type

ReferencePoints{RD1, RD3}

Log of reference pressures, reference temperatures, and volume mixing ratios used by the lookup table

RRTMGP.LookUpTables.LookUpPlanck — Type

LookUpPlanck{PD1, PD2, PD4}

Look up data for Planck source calculations.

RRTMGP.LookUpTables.BandData — Type

BandData{BNDI1, BNDI2, BNDD2}

Band/g-point data for the lookup table.

RRTMGP.LookUpTables.LookUpLW — Type

LookUpLW{FT, IA3D, FTA4D, BND, P, R, LMNR} <: AbstractLookUp

Longwave lookup tables, used to compute optical properties.

Fields

idx_h2o: vmr array index for h2o

p_ref_tropo: Reference pressure separating upper and lower atmosphere

p_ref_min: minimum pressure supported by RRTMGP lookup tables

key_species: major absorbing species in each band (2, n_atmos_layers, n_bnd)

kmajor: major absorption coefficient (n_η, n_p_ref + 1, n_t_ref, n_gpt)

planck: lookup data for Planck source calculations

band_data: band data

ref_points: reference temperatures, pressures and volume mixing ratios

minor_lower: lookup data for minor gases in the lower atmosphere

minor_upper: lookup data for minor gases in the upper atmosphere

RRTMGP.LookUpTables.LookUpSW — Type

LookUpSW{FT, IA3D, FTA1D, FTA3D, FTA4D, BND, R, LMNR} <: AbstractLookUp

Shortwave lookup tables, used to compute optical properties.

Fields

idx_h2o: vmr array index for h2o

p_ref_tropo: Reference pressure separating upper and lower atmosphere

p_ref_min: minimum pressure supported by RRTMGP lookup tables

solar_src_tot: total solar irradiation

key_species: major absorbing species in each band (2, n_atmos_layers, n_bnd)

kmajor: major absorption coefficient (n_η, n_p_ref + 1, n_t_ref, n_gpt)

band_data: band data

ref_points: reference temperatures, pressures and volume mixing ratios

rayl_lower: Rayleigh absorption coefficient for lower atmosphere (n_η, n_t_ref, n_gpt)

rayl_upper: Rayleigh absorption coefficient for upper atmosphere (n_η, n_t_ref, n_gpt)

solar_src_scaled: relative solar source contribution from each g-point (n_gpt)

minor_lower: lookup data for minor gases in the lower atmosphere

minor_upper: lookup data for minor gases in the upper atmosphere

RRTMGP.LookUpTables.LookUpCld — Type

LookUpCld{D, B, L, I, W} <: AbstractLookUp

Lookup table for cloud optics.

This struct stores the lookup tables for determing extinction coeffient, single-scattering albedo, and asymmetry parameter g as a function of effective radius. We compute the optical depth tau (=exintinction coeff * condensed water path) and the products taussa and taussa*g for liquid and ice cloud separately. These are used to determine the optical properties of ice and water cloud together.

Fields

dims: nband, nrghice, nsize_liq, nsize_ice, pair

bounds: particle size lower and upper bounds and factor for LUT interpolation for liquid and ice particles

liqdata: liquid extinction coefficient, single scattering albedo and symmetry paramter (3*nsize_liq, nbnd)

icedata: ice extinction coefficient, single scattering albedo and symmetry paramter (3*nsize_ice, nbnd, nrghice)

bnd_lims_wn: beginning and ending wavenumber for each band (2, nband) cm⁻¹

RRTMGP.LookUpTables.PadeCld — Type

PadeCld{FTA1D, FTA2D, FTA3D, FTA4D} <: AbstractLookUp

Pade coefficients for cloud optics.

This struct stores the Pade coefficients for determing extinction coeffient, single-scattering albedo, and asymmetry parameter g as a function of effective radius. We compute the optical depth tau (=exintinction coeff * condensed water path) and the products taussa and taussa*g for liquid and ice cloud separately. These are used to determine the optical properties of ice and water cloud together.

Fields

pade_extliq: pade coefficients for liquid extinction (nband, nsizereg, ncoeff_ext)

pade_ssaliq: pade coefficients for liquid single scattening albedo (nband, nsizereg, ncoeff_ssa_g)

pade_asyliq: pade coefficients for liquid asymmetry paramter (nband, nsizereg, ncoeff_ssa_g)

pade_extice: pade coefficients for ice extinction (nband, nsizereg, ncoeff_ext, nrghice)

pade_ssaice: pade coefficients for ice single scattening albedo (nband, nsizereg, ncoeff_ssa_g, nrghice)

pade_asyice: pade coefficients for ice asymmetry paramter (nband, nsizereg, ncoeff_ssa_g, nrghice)

pade_sizreg_extliq: pade size regime boundaries for liquid extinction coefficient for pade interpolation (nbound) μm

pade_sizreg_ssaliq: pade size regime boundaries for liquid single scattering albedo for pade interpolation (nbound) μm

pade_sizreg_asyliq: pade size regime boundaries for liquid asymmetry parameter for pade interpolation (nbound) μm

pade_sizreg_extice: pade size regime boundaries for ice extinction coefficient for pade interpolation (nbound) μm

pade_sizreg_ssaice: pade size regime boundaries for ice single scattering albedo for pade interpolation (nbound) μm

pade_sizreg_asyice: pade size regime boundaries for ice asymmetry parameter for pade interpolation (nbound) μm

bnd_lims_wn: beginning and ending wavenumber for each band (2, nband) cm⁻¹

RRTMGP.LookUpTables.AbstractLookUp — Type

AbstractLookUp

Abstract lookup table for longwave and shortwave problems.