Lookup tables

RRTMGP.LookUpTables.LookUpLW — Type

LookUpLW{I,FT,UI8,UI8A1D,IA1D,IA2D,IA3D,FTA1D,FTA2D,FTA3D,FTA4D} <: 
    AbstractLookUp{I,FT}

Longwave lookup tables, used to compute optical properties.

Fields

n_gases: number of gases used in the lookup table

n_bnd: number of longwave bands

n_gpt: number of g-points

n_atmos_layers: number of atmospheric layers (=2, lower and upper atmospheres)

n_t_ref: number of reference temperatures for absorption coefficient lookup table

n_p_ref: number of reference pressures for absorption lookup table

n_η: number of reference binary mixing fractions, for absorption coefficient lookup table

n_t_plnk: number of reference temperatures, for Planck source calculations

n_maj_absrb: number of major absorbing gases

n_min_absrb: number of minor absorbing gases

n_min_absrb_lower: number of minor absorbers in lower atmosphere

n_min_absrb_upper: number of minor absorbers in upper atmosphere

n_absrb_ext

n_contrib_lower: number of minor contributors in the lower atmosphere

n_contrib_upper: number of minor contributors in the upper atmosphere

idx_h2o: vmr array index for h2o

p_ref_tropo: Reference pressure separating upper and lower atmosphere

t_ref_absrb: Reference temperature

p_ref_absrb: Reference pressure

p_ref_min: minimum pressure supported by RRTMGP lookup tables

Δ_t_ref: Δt for reference temperature values (Δt is constant)

Δ_ln_p_ref: Δ for log of reference pressure values (Δp is constant)

idx_gases_minor_lower: indices for minor gases contributing to absorption in the lower atmosphere (n_min_absrb_lower)

idx_gases_minor_upper: indices for minor gases contributing to absorption in the upper atmosphere (n_min_absrb_upper)

idx_scaling_gas_lower: indices for scaling gases in the lower atmosphere (n_min_absrb_lower)

idx_scaling_gas_upper: indices for scaling gases in the upper atmosphere (n_min_absrb_upper)

key_species: major absorbing species in each band (2, n_atmos_layers, n_bnd)

kmajor: major absorption coefficient (n_gpt, n_η, n_p_ref, n_t_ref)

kminor_lower: minor absorption coefficient in lower atmosphere (n_contrib_lower, n_η, n_t_ref)

kminor_upper: minor absorption coefficient in upper atmosphere (n_contrib_upper, n_η, n_t_ref)

kminor_start_lower

kminor_start_upper

planck_fraction: Planck fraction (n_gpt, n_η, n_p_ref, n_t_ref)

t_planck: reference temperatures for Planck source calculations (n_t_plnk)

totplnk: total Planck source for each band (n_t_plnk, n_bnd)

major_gpt2bnd: map from g-point to band

bnd_lims_gpt: starting and ending g-point for each band (2, n_bnd)

bnd_lims_wn: starting and ending wavenumber for each band (2, n_bnd)

minor_lower_gpt_lims: g-point limits for minor contributors in lower atmosphere (2, n_contrib_lower)

minor_upper_gpt_lims: g-point limits for minor contributors in upper atmosphere (2, n_contrib_upper)

minor_lower_bnd: band number for minor contributor in the lower atmosphere (n_contrib_lower)

minor_upper_bnd: band number for minor contributor in the upper atmosphere (n_contrib_upper)

minor_lower_bnd_st: starting index to idx_gases_minor_lower for each band (n_bnd)

minor_upper_bnd_st: starting index to idx_gases_minor_upper for each band (n_bnd)

minor_lower_gpt_sh: shift in kminor_lower for each band (n_min_absrb_lower)

minor_upper_gpt_sh: shift in kminor_upper for each band (n_min_absrb_upper)

minor_lower_scales_with_density: minor gas (lower atmosphere) scales with density? (n_min_absrb_lower)

minor_upper_scales_with_density: minor gas (upper atmosphere) scales with density? (n_min_absrb_upper)

lower_scale_by_complement: minor gas (lower atmosphere) scales by compliment (n_min_absrb_lower)

upper_scale_by_complement: minor gas (upper atmosphere) scales by compliment (n_min_absrb_upper)

p_ref: reference pressures used by the lookup table (n_p_ref)

t_ref: reference temperatures used by the lookup table (n_t_ref)

vmr_ref: reference volume mixing ratios used by the lookup table (2, n_gases, n_t_ref)

RRTMGP.LookUpTables.LookUpSW — Type

LookUpSW{I,FT,UI8,UI8A1D,IA1D,IA2D,IA3D,FTA1D,FTA2D,FTA3D,FTA4D} <:
    AbstractLookUp{I,FT}

Shortwave lookup tables, used to compute optical properties.

Fields

n_gases: number of gases used in the lookup table

n_bnd: number of shortwave bands

n_gpt: number of g-points

n_atmos_layers: number of atmospheric layers (=2, lower and upper atmospheres)

n_t_ref: number of reference temperatures for absorption coefficient lookup table

n_p_ref: number of reference pressures for absorption lookup table

n_η: number of reference binary mixing fractions, for absorption coefficient lookup table

n_maj_absrb: number of major absorbing gases

n_min_absrb: number of minor absorbing gases

n_min_absrb_lower: number of minor absorbers in lower atmosphere

n_min_absrb_upper: number of minor absorbers in upper atmosphere

n_absrb_ext

n_contrib_lower: number of minor contributors in the lower atmosphere

n_contrib_upper: number of minor contributors in the upper atmosphere

idx_h2o: vmr array index for h2o

p_ref_tropo: Reference pressure separating upper and lower atmosphere

t_ref_absrb: Reference temperature

p_ref_absrb: Reference pressure

p_ref_min: minimum pressure supported by RRTMGP lookup tables

Δ_t_ref: Δt for reference temperature values (Δt is constant, array is linearly equispaced)

Δ_ln_p_ref: Δ for log of reference pressure values (Δ is constant, equispaced on logarithmic scale)

solar_src_tot: total solar irradiation

idx_gases_minor_lower: indices for minor gases contributing to absorption in the lower atmosphere (n_min_absrb_lower)

idx_gases_minor_upper: indices for minor gases contributing to absorption in the upper atmosphere (n_min_absrb_upper)

idx_scaling_gas_lower: indices for scaling gases in the lower atmosphere (n_min_absrb_lower)

idx_scaling_gas_upper: indices for scaling gases in the upper atmosphere (n_min_absrb_upper)

key_species: major absorbing species in each band (2, n_atmos_layers, n_bnd)

kmajor: major absorption coefficient (n_gpt, n_η, n_p_ref, n_t_ref)

kminor_lower: minor absorption coefficient in lower atmosphere (n_contrib_lower, n_η, n_t_ref)

kminor_upper: minor absorption coefficient in upper atmosphere (n_contrib_upper, n_η, n_t_ref)

kminor_start_lower

kminor_start_upper

major_gpt2bnd: map from g-point to band

bnd_lims_gpt: starting and ending g-point for each band (2, n_bnd)

bnd_lims_wn: starting and ending wavenumber for each band (2, n_bnd)

minor_lower_gpt_lims: g-point limits for minor contributors in lower atmosphere (2, n_contrib_lower)

minor_upper_gpt_lims: g-point limits for minor contributors in upper atmosphere (2, n_contrib_upper)

minor_lower_bnd: band number for minor contributor in the lower atmosphere (n_contrib_lower)

minor_upper_bnd: band number for minor contributor in the upper atmosphere (n_contrib_upper)

minor_lower_bnd_st: starting index to idx_gases_minor_lower for each band

minor_upper_bnd_st: starting index to idx_gases_minor_upper for each band

minor_lower_gpt_sh: shift in kminor_lower for each band (n_min_absrb_lower)

minor_upper_gpt_sh: shift in kminor_upper for each band (n_min_absrb_upper)

minor_lower_scales_with_density: minor gas (lower atmosphere) scales with density? (n_min_absrb_lower)

minor_upper_scales_with_density: minor gas (upper atmosphere) scales with density? (n_min_absrb_upper)

lower_scale_by_complement: minor gas (lower atmosphere) scales by compliment (n_min_absrb_lower)

upper_scale_by_complement: minor gas (upper atmosphere) scales by compliment (n_min_absrb_upper)

p_ref: reference pressures used by the lookup table (n_p_ref)

t_ref: reference temperatures used by the lookup table (n_t_ref)

vmr_ref: reference volume mixing ratios used by the lookup table (2, n_gases, n_t_ref)

rayl_lower: Rayleigh absorption coefficient for lower atmosphere (n_gpt, n_η, n_t_ref)

rayl_upper: Rayleigh absorption coefficient for upper atmosphere (n_gpt, n_η, n_t_ref)

solar_src_scaled: relative solar source contribution from each g-point (n_gpt)

RRTMGP.LookUpTables.LookUpCld — Type

LookUpCld{I,FT,FTA1D,FTA2D,FTA3D,FTA4D}

Lookup table for cloud optics.

This struct stores the tables and Pade coefficients for determing extinction coeffient, single-scattering albedo, and asymmetry parameter g as a function of effective radius. We compute the optical depth tau (=exintinction coeff * condensed water path) and the products taussa and taussa*g for liquid and ice cloud separately. These are used to determine the optical properties of ice and water cloud together.

Fields

nband: number of bands

nrghice: number of ice roughness types

nsize_liq: number of liquid particle sizes

nsize_ice: number of ice particle sizes

nsizereg: number of size regimes

ncoeff_ext: number of extinction coefficients for pade approximation

ncoeff_ssa_g: number of ssa/g coefficients for pade approximation

nbound: number of size regime boundaries for pade interpolation

pair: pair = 2

radliq_lwr: liquid particle size lower bound for LUT interpolation

radliq_upr: liquid particle size upper bound for LUT interpolation

radliq_fac: factor for calculating LUT interpolation for liquid particle

radice_lwr: ice particle size lower bound for LUT interpolation

radice_upr: ice particle size upper bound for LUT interpolation

radice_fac: factor for calculating LUT interpolation for ice particle

lut_extliq: LUT liquid extinction coefficient (nsize_liq, nbnd) m²/g

lut_ssaliq: LUT liquid single scattering albedo (nsize_liq, nbnd)

lut_asyliq: LUT liquid asymmetry parameter (nsize_liq, nbnd)

lut_extice: LUT ice extinction coefficient (nsize_ice, nband, nrghice) m²/g

lut_ssaice: LUT ice single scattering albedo (nsize_ice, nband, nrghice)

lut_asyice: LUT ice asymmetry parameter (nsize_ice, nband, nrghice)

pade_extliq: pade coefficients for liquid extinction (nband, nsizereg, ncoeff_ext)

pade_ssaliq: pade coefficients for liquid single scattening albedo (nband, nsizereg, ncoeff_ssa_g)

pade_asyliq: pade coefficients for liquid asymmetry paramter (nband, nsizereg, ncoeff_ssa_g)

pade_extice: pade coefficients for ice extinction (nband, nsizereg, ncoeff_ext, nrghice)

pade_ssaice: pade coefficients for ice single scattening albedo (nband, nsizereg, ncoeff_ssa_g, nrghice)

pade_asyice: pade coefficients for ice asymmetry paramter (nband, nsizereg, ncoeff_ssa_g, nrghice)

pade_sizreg_extliq: pade size regime boundaries for liquid extinction coefficient for pade interpolation (nbound) μm

pade_sizreg_ssaliq: pade size regime boundaries for liquid single scattering albedo for pade interpolation (nbound) μm

pade_sizreg_asyliq: pade size regime boundaries for liquid asymmetry parameter for pade interpolation (nbound) μm

pade_sizreg_extice: pade size regime boundaries for ice extinction coefficient for pade interpolation (nbound) μm

pade_sizreg_ssaice: pade size regime boundaries for ice single scattering albedo for pade interpolation (nbound) μm

pade_sizreg_asyice: pade size regime boundaries for ice asymmetry parameter for pade interpolation (nbound) μm

bnd_lims_wn: beginning and ending wavenumber for each band (2, nband) cm⁻¹

use_lut: use LUT (default) or pade interpolation

RRTMGP.LookUpTables.AbstractLookUp — Type

AbstractLookUp{I,FT}

Abstract lookup table for longwave and shortwave problems.