Optics

RRTMGP.Optics.AbstractOpticalProps — Type

AbstractOpticalProps{FT,FTA2D}

Optical properties for one scalar and two stream calculations.

RRTMGP.Optics.OneScalar — Type

OneScalar{FT,FTA1D,FTA2D,I,AD} <: AbstractOpticalProps{FT,FTA2D}

Single scalar approximation for optical depth, used in calculations accounting for extinction and emission

Fields

τ

Optical Depth

angle_disc

Angular discretization

RRTMGP.Optics.TwoStream — Type

TwoStream{FT,FTA2D} <: AbstractOpticalProps{FT,FTA2D}

Two stream approximation for optical properties, used in calculations accounting for extinction and emission

Fields

τ

Optical depth

ssa

Single-scattering albedo

g

Asymmetry parameter

RRTMGP.Optics.compute_col_dry! — Function

compute_col_dry!(
    p_lev,
    col_dry,
    param_set,
    vmr_h2o,
    lat,
    max_threads = Int(256),
)

This function computes the column amounts of dry or moist air.

RRTMGP.Optics.compute_optical_props! — Function

compute_optical_props!(
    op::AbstractOpticalProps{FT},
    as::AtmosphericState{FT},
    sf::AbstractSourceLW{FT},
    gcol::I,
    igpt::I,
    lkp::LookUpLW{I,FT},
    lkp_cld::Union{LookUpCld,Nothing} = nothing,
) where {I<:Int,FT<:AbstractFloat}

Computes optical properties for the longwave problem.

compute_optical_props!(
    op::AbstractOpticalProps{FT},
    as::AtmosphericState{FT},
    gcol::I,
    igpt::I,
    lkp::LookUpSW{I,FT},
    lkp_cld::Union{LookUpCld,Nothing} = nothing,
) where {I<:Int,FT<:AbstractFloat}

Computes optical properties for the shortwave problem.

compute_optical_props!(
    op::AbstractOpticalProps{FT},
    as::GrayAtmosphericState{FT},
    sf::AbstractSourceLW{FT},
    gcol::I,
    igpt::I = 1,
) where {FT<:AbstractFloat}

Computes optical properties for the longwave gray radiation problem.

compute_optical_props!(
    op::AbstractOpticalProps{FT},
    as::GrayAtmosphericState{FT},
    gcol::I,
    igpt::I = 1,
) where {FT<:AbstractFloat}

Computes optical properties for the shortwave gray radiation problem.

RRTMGP.Optics.compute_col_dry_kernel! — Function

compute_col_dry_kernel!(
    col_dry,
    p_lev,
    mol_m_dry,
    mol_m_h2o,
    avogadro
    helmert1,
    vmr_h2o
    lat,
    glaycol,
)

This function computes the column amounts of dry or moist air.

RRTMGP.Optics.compute_interp_fractions — Function

compute_interp_fractions(
    lkp::AbstractLookUp{I,FT},
    vmr,
    p_lay,
    t_lay,
    tropo,
    ibnd,
    glaycol,
) where {I<:Int,FT<:AbstractFloat}

compute interpolation fractions for binary species parameter, pressure and temperature.

RRTMGP.Optics.compute_interp_frac_temp — Function

compute_interp_frac_temp(
    lkp::AbstractLookUp{I,FT},
    t_lay,
    glay,
    gcol,
) where {I<:Int,FT<:AbstractFloat}

compute interpolation fraction for temperature.

RRTMGP.Optics.compute_interp_frac_press — Function

compute_interp_frac_press(
    lkp::AbstractLookUp,
    p_lay,
    tropo,
    glay,
    gcol,
)

Compute interpolation fraction for pressure.

RRTMGP.Optics.compute_interp_frac_η — Function

compute_interp_frac_η(
    lkp::AbstractLookUp{I,FT},
    vmr,
    tropo,
    jtemp,
    ibnd,
    glay,
    gcol,
) where {FT<:AbstractFloat,I<:Int}

Compute interpolation fraction for binary species parameter.

RRTMGP.Optics.compute_τ_ssa_lw_src! — Function

compute_τ_ssa_lw_src!(
    lkp::AbstractLookUp{I,FT},
    vmr,
    col_dry,
    igpt,
    ibnd,
    p_lay::FT,
    t_lay,
    glaycol,
    src_args...,
) where {FT<:AbstractFloat,I<:Int}

Compute optical thickness, single scattering albedo, asymmetry parameter and longwave sources whenever applicable.

RRTMGP.Optics.compute_τ_minor — Function

compute_τ_minor(
    lkp::AbstractLookUp,
    tropo::I,
    vmr,
    vmr_h2o::FT,
    col_dry,
    p_lay::FT,
    t_lay::FT,
    jtemp::I,
    ftemp::FT,
    jη1::I,
    jη2::I,
    fη1::FT,
    fη2::FT,
    igpt,
    ibnd,
    glay,
    gcol,
) where {FT<:AbstractFloat,I<:Int}

Compute optical thickness contributions from minor gases.

RRTMGP.Optics.compute_τ_rayleigh — Function

compute_τ_rayleigh(
    lkp::LookUpSW,
    tropo::I,
    col_dry::FT,
    vmr_h2o::FT,
    jtemp::I,
    ftemp::FT,
    jη1::I,
    jη2::I,
    fη1::FT,
    fη2::FT,
    igpt::I,
) where {FT<:AbstractFloat,I<:Int}

Compute Rayleigh scattering optical depths for shortwave problem

RRTMGP.Optics.compute_lw_planck_src! — Function

compute_lw_planck_src!(
    lkp::LookUpLW,
    jη1,
    jη2,
    fη1,
    fη2,
    jpresst,
    fpress,
    jtemp,
    ftemp,
    t_lay,
    igpt,
    ibnd,
    glay,
    gcol,
    sf,
    t_lev,
    t_sfc,
)

Computes Planck sources for the longwave problem.

RRTMGP.Optics.add_cloud_optics_2stream — Function

add_cloud_optics_2stream(
    op::TwoStream,
    as::AtmosphericState,
    lkp::LookUpLW,
    lkp_cld,
    glaycol,
    ibnd,
    igpt,
)

This function computes the longwave TwoStream clouds optics properties and adds them to the TwoStream longwave gas optics properties.

add_cloud_optics_2stream(
    op::TwoStream,
    as::AtmosphericState,
    lkp::LookUpSW,
    lkp_cld,
    glaycol,
    ibnd,
    igpt,
)

This function computes the longwave TwoStream clouds optics properties and adds them to the TwoStream shortwave gase optics properties.

add_cloud_optics_2stream(op::OneScalar, args...)

Cloud optics is currently only supported for the TwoStream solver.

RRTMGP.Optics.compute_cld_props — Function

compute_cld_props(lkp_cld, as, glay, gcol, ibnd, igpt)

This function computed the TwoSteam cloud optics properties using either the lookup table method or pade method.

RRTMGP.Optics.pade_eval — Function

pade_eval(
    ibnd,
    re,
    irad,
    m,
    n,
    pade_coeffs,
    irgh::Union{Int,Nothing} = nothing,
)

Evaluate Pade approximant of order [m/n]

RRTMGP.Optics.compute_optical_props_kernel! — Function

compute_optical_props_kernel!(
    op::AbstractOpticalProps{FT},
    as::GrayAtmosphericState{FT},
    glaycol,
    source::AbstractSourceLW{FT},
) where {FT<:AbstractFloat}

This function computes the optical properties using the gray atmosphere assumption for the longwave solver.

compute_optical_props_kernel!(
    op::AbstractOpticalProps{FT},
    as::GrayAtmosphericState{FT},
    glaycol,
) where {FT<:AbstractFloat}

This function computes the optical properties using the gray atmosphere assumption for the shortwave solver.

RRTMGP.Optics.compute_sources_gray_kernel! — Function

compute_sources_gray_kernel!(
    source::AbstractSourceLW{FT},
    as::GrayAtmosphericState{FT},
    glaycol,
) where {FT<:AbstractFloat}

This function computes the Planck sources for the gray longwave solver.

RRTMGP.Optics.interp1d — Function

interp1d(xi::FT, x, y, col) where {FT<:AbstractFloat}

perform 1D linear interpolation.

RRTMGP.Optics.interp2d — Function

interp2d(
    fη1::FT,
    fη2::FT,
    ftemp::FT,
    coeff::FTA3D,
    igpt::I,
    jη1::I,
    jη2::I,
    jtemp::I,
) where {FT<:AbstractFloat,FTA3D<:AbstractArray{FT,3},I<:Int}

Perform 2D linear interpolation.

fminor[1, 1] = (FT(1) - fη1) * (1-ftemp)

fminor[2, 1] = fη1 * (1-ftemp)

fminor[1, 2] = (FT(1) - fη2) * ftemp

fminor[2, 2] = fη2 * ftemp

RRTMGP.Optics.interp3d — Function

interp3d(
    jη1::I,
    jη2::I,
    fη1::FT,
    fη2::FT,
    jtemp::I,
    ftemp::FT,
    jpresst::I,
    fpress::FT,
    coeff::FTA4D,
    igpt::I,
    s1::FT = FT(1),
    s2::FT = FT(1),
) where {FT<:AbstractFloat,FTA4D<:AbstractArray{FT,4},I<:Int}

Perform 3D linear interpolation.

fmajor[1, 1, itemp] = (FT(1) - fpress) * fminor[1, itemp]

fmajor[2, 1, itemp] = (FT(1) - fpress) * fminor[2, itemp]

fmajor[1, 2, itemp] = fpress * fminor[1, itemp]

fmajor[2, 2, itemp] = fpress * fminor[2, itemp]

where,

fminor[1, itemp=1] = (FT(1) - fη1) * (1-ftemp)

fminor[2, itemp=1] = fη1 * (1-ftemp)

fminor[1, itemp=2] = (FT(1) - fη2) * ftemp

fminor[2, itemp=2] = fη2 * ftemp

RRTMGP.Optics.increment! — Function

increment!(
    op::TwoStream{FT},
    τ2,
    ssa2,
    g2,
    glay,
    gcol,
    igpt,
) where {FT<:AbstractFloat}

Increment TwoStream optical properties op for layer glay and column gcol with optical thickness τ2, single scattering albedo ssa2 and symmetry parameter g2.

RRTMGP.Optics.delta_scale — Function

delta_scale(τ, ssa, g)

delta-scale two stream optical properties.