RTE Solver

RRTMGP.RTESolver.solve_lw! — Function

solve_lw!(
    slv::Solver,
    max_threads::Int,
    lkp_args...
)

Solver for the longwave radiation problem

RRTMGP.RTESolver.rte_lw_noscat_solve! — Function

rte_lw_noscat_solve!(
    flux::FluxLW{FT},
    flux_lw::FluxLW{FT},
    src_lw::SourceLWNoScat{FT},
    bcs_lw::LwBCs{FT},
    op::OneScalar{FT},
    nlay,
    ncol,
    major_gpt2bnd::AbstractArray{UInt8,1},
    max_threads,
    as::AbstractAtmosphericState{FT},
    lkp_args...,
) where {FT<:AbstractFloat}

No scattering solver for the longwave problem. LW fluxes, no scattering, mu (cosine of integration angle) specified by column Does radiation calculation at user-supplied angles; converts radiances to flux using user-supplied weights. Currently, only ngaussangles = 1 supported.

RRTMGP.RTESolver.rte_lw_2stream_solve! — Function

rte_lw_2stream_solve!(
    flux::FluxLW{FT},
    flux_lw::FluxLW{FT},
    src_lw::SourceLW2Str{FT},
    bcs_lw::LwBCs{FT},
    op::TwoStream{FT},
    nlay,
    ncol,
    major_gpt2bnd::AbstractArray{UInt8,1},
    max_threads,
    as::AbstractAtmosphericState{FT},
    lkp_args...,
) where {FT<:AbstractFloat}

Two stream solver for the longwave problem.

RRTMGP provides source functions at each level using the spectral mapping of each adjacent layer. Combine these for two-stream calculations lw combine sources (rtelw2streamcombinesources!) combine RRTMGP-specific sources at levels compute layer reflectance, transmittance compute total source function at levels using linear-in-tau (rtelw2streamsource!) transport (addinglw!)

RRTMGP.RTESolver.solve_sw! — Function

solve_sw!(
    slv::Solver,
    max_threads::Int,
    lkp_args...,
)

Solver for the shortwave radiation problem

RRTMGP.RTESolver.rte_sw_noscat_solve! — Function

rte_sw_noscat_solve!(
    flux::FluxSW{FT},
    flux_sw::FluxSW{FT},
    op::OneScalar{FT},
    bcs_sw::SwBCs{FT},
    nlay,
    ncol,
    solar_src_scaled::AbstractArray{FT,1},
    max_threads,
    as::AbstractAtmosphericState{FT},
    lkp_args...,
) where {FT<:AbstractFloat}

No-scattering solver for the shortwave problem. (Extinction-only i.e. solar direct beam)

RRTMGP.RTESolver.rte_sw_2stream_solve! — Function

rte_sw_2stream_solve!(
    flux::FluxSW{FT},
    flux_sw::FluxSW{FT},
    op::TwoStream{FT},
    bcs_sw::SwBCs{FT},
    src_sw::SourceSW2Str{FT},
    nlay,
    ncol,
    major_gpt2bnd::AbstractArray{UInt8,1},
    solar_src_scaled::AbstractArray{FT,1},
    max_threads,
    as::AbstractAtmosphericState{FT},
    lkp_args...,
) where {FT<:AbstractFloat}

Two stream solver for the shortwave problem.

RRTMGP.RTESolver.rte_lw_noscat_source! — Function

rte_lw_noscat_source!(
    src_lw::SourceLWNoScat{FT},
    op::OneScalar{FT},
    gcol,
    nlay,
) where {FT<:AbstractFloat}

Compute LW source function for upward and downward emission at levels using linear-in-tau assumption See Clough et al., 1992, doi: 10.1029/92JD01419, Eq 13

RRTMGP.RTESolver.rte_lw_noscat_transport! — Function

rte_lw_noscat_transport!(
    src_lw::SourceLWNoScat{FT},
    bcs_lw::LwBCs{FT},
    op::OneScalar{FT},
    gcol,
    flux::FluxLW{FT},
    igpt,
    ibnd,
    nlay,
    nlev,
) where {FT<:AbstractFloat}

Transport for no-scattering longwave problem.

RRTMGP.RTESolver.rte_lw_2stream_combine_sources! — Function

rte_lw_2stream_combine_sources!(src::SourceLW2Str, gcol, nlev, ncol)

RRTMGP provides source functions at each level using the spectral mapping of each adjacent layer. This function combines these for two-stream calculations.

RRTMGP.RTESolver.rte_lw_2stream_source! — Function

rte_lw_2stream_source!(
    op::TwoStream{FT},
    src_lw::SourceLW2Str{FT},
    gcol,
    nlay,
    ncol,
) where {FT<:AbstractFloat}

This function combines RRTMGP-specific sources at levels, computes layer reflectance, transmittance, and total source function at levels using linear-in-tau approximation.

RRTMGP.RTESolver.adding_lw! — Function

adding_lw!(
    flux::FluxLW{FT},
    src_lw::SL,
    bcs_lw::BCL,
    gcol,
    igpt,
    ibnd,
    nlev,
    ncol,
) where {FT<:AbstractFloat,SL<:SourceLW2Str{FT},BCL<:LwBCs{FT}}

Transport of diffuse radiation through a vertically layered atmosphere, for the longwave problem. Equations are after Shonk and Hogan 2008, doi:10.1175/2007JCLI1940.1 (SH08)

RRTMGP.RTESolver.rte_sw_noscat_solve_kernel! — Function

rte_sw_noscat_solve_kernel!(
    flux::FluxSW{FT},
    op::OneScalar{FT},
    bcs_sw::SwBCs{FT},
    solar_frac::FT,
    gcol,
    nlev,
) where {FT<:AbstractFloat}

No-scattering solver for the shortwave problem. (Extinction-only i.e. solar direct beam)

RRTMGP.RTESolver.sw_two_stream! — Function

sw_two_stream!(
    op::TwoStream{FT},
    src_sw::SourceSW2Str{FT},
    bcs_sw::SwBCs{FT},
    gcol,
    nlay,
) where {FT<:AbstractFloat}

Computes cell properties (transmittance and reflectance) for direct and diffuse radiation

Two-stream solutions to direct and diffuse reflectance and transmittance for a layer with optical depth tau, single scattering albedo w0, and asymmetery parameter g.

Equations are developed in Meador and Weaver, 1980, doi:10.1175/1520-0469(1980)037<0630:TSATRT>2.0.CO;2

RRTMGP.RTESolver.sw_source_2str! — Function

sw_source_2str!(
    src_sw::SourceSW2Str{FT},
    bcs_sw::SwBCs{FT},
    gcol,
    flux::FluxSW{FT},
    solar_frac::FT,
    ibnd,
    nlay,
) where {FT<:AbstractFloat}

Direct-beam and source for diffuse radiation

RRTMGP.RTESolver.adding_sw! — Function

adding_sw!(
    src_sw::SourceSW2Str{FT},
    bcs_sw::SwBCs{FT},
    gcol,
    flux::FluxSW{FT},
    igpt,
    ibnd,
    nlev,
) where {FT<:AbstractFloat}

Transport for the two stream shortwave problem.

Transport of diffuse radiation through a vertically layered atmosphere. Equations are after Shonk and Hogan 2008, doi:10.1175/2007JCLI1940.1 (SH08)